refactor: reorganize repository by splitting up files
New files have been made for each class and been sorted into subdirectories.

While many of the files have been changed, the overall compile library and binary should be virtually the same
