+++ include/chemparse/util/periodictable.hpp
