+++ src/chemparse/util/periodictable.cpp
@@ -1,4 +1,4 @@
-#include "periodictable.hpp"
+#include "chemparse/util/periodictable.hpp"
 
 unsigned char periodic_table[] = {
     0x7b, 0x0a, 0x20, 0x20, 0x20, 0x20, 0x22, 0x65, 0x6c, 0x65, 0x6d, 0x65,
